Fix int32 overflow in legacy directory (1 file)

+++ paddle/phi/kernels/legacy/gpu/legacy_generate_proposals_kernel.cu
@@ -44,8 +44,8 @@ static std::pair<phi::DenseTensor, phi::DenseTensor> ProposalForOneImage(
   // 1. pre nms
   phi::DenseTensor scores_sort, index_sort;
   phi::funcs::SortDescending<T>(dev_ctx, scores, &scores_sort, &index_sort);
-  int num = scores.numel();
-  int pre_nms_num = (pre_nms_top_n <= 0 || pre_nms_top_n > num) ? scores.numel()
+  int64_t num = scores.numel();
+  int64_t pre_nms_num = (pre_nms_top_n <= 0 || pre_nms_top_n > num) ? scores.numel()
                                                                 : pre_nms_top_n;
   scores_sort.Resize({pre_nms_num, 1});
   index_sort.Resize({pre_nms_num, 1});
